VB-5218, move confirm update to own controller
