+++ server/routes/visit.test.ts
@@ -337,72 +337,4 @@ describe('/visit/:reference', () => {
         })
     })
   })
-
-  describe('GET /visit/:reference/update/confirm-update', () => {
-    it('should render the confirm update page', () => {
-      app = appWithAllRoutes({
-        services: {
-          auditService,
-          prisonerSearchService,
-          prisonerVisitorsService,
-          supportedPrisonsService,
-          visitService,
-          visitSessionsService,
-        },
-        sessionData: {
-          selectedEstablishment: { ...prison, policyNoticeDaysMin: 4 },
-        } as SessionData,
-      })
-
-      return request(app)
-        .get(`/visit/${visit.reference}/update/confirm-update`)
-        .expect(200)
-        .expect('Content-Type', /html/)
-        .expect(res => {
-          const $ = cheerio.load(res.text)
-          expect($('.govuk-back-link').attr('href')).toBe(`/visit/${visit.reference}`)
-          expect($('h1').text().trim()).toContain('This visit is in less than 4 days.')
-          expect($('h1').text().trim()).toContain('Do you want to update the booking?')
-          expect($('form').attr('action')).toBe('/visit/ab-cd-ef-gh/update/confirm-update')
-        })
-    })
-  })
-
-  describe('POST /visit/:reference/update/confirm-update', () => {
-    it('should redirect back to the visit summary if choosing not to proceed with update', () => {
-      return request(app)
-        .post('/visit/ab-cd-ef-gh/update/confirm-update')
-        .send('confirmUpdate=no')
-        .expect(302)
-        .expect('location', '/visit/ab-cd-ef-gh')
-        .expect(res => {
-          expect(visitSessionData).not.toHaveProperty('overrideBookingWindow')
-        })
-    })
-
-    it('should redirect to select visitors page if choosing to proceed with update', () => {
-      return request(app)
-        .post('/visit/ab-cd-ef-gh/update/confirm-update')
-        .send('confirmUpdate=yes')
-        .expect(302)
-        .expect('location', '/visit/ab-cd-ef-gh/update/select-visitors')
-        .expect(res => {
-          expect(visitSessionData.overrideBookingWindow).toBe(true)
-        })
-    })
-
-    it('should should redirect to confirm update page with errors set if no option selected', () => {
-      return request(app)
-        .post('/visit/ab-cd-ef-gh/update/confirm-update')
-        .send('confirmUpdate=')
-        .expect(302)
-        .expect('location', '/visit/ab-cd-ef-gh/update/confirm-update')
-        .expect(() => {
-          expect(visitSessionData).not.toHaveProperty('overrideBookingWindow')
-          expect(flashProvider).toHaveBeenCalledWith('errors', [
-            { location: 'body', msg: 'No option selected', path: 'confirmUpdate', type: 'field' },
-          ])
-        })
-    })
-  })
 })
