+++ server/routes/visit/index.ts
@@ -7,11 +7,13 @@ import VisitDetailsController from './visitDetailsController'
 import CancelVisitController from './cancelVisitController'
 import { isValidVisitReference } from '../validationChecks'
 import ClearNotificationsController from './clearNotificationsController'
+import ConfirmUpdateController from './confirmUpdateController'
 
 export default function routes(services: Services): Router {
   const router = Router()
 
   const get = (path: string | string[], handler: RequestHandler) => router.get(path, asyncMiddleware(handler))
+  const post = (path: string, handler: RequestHandler) => router.post(path, asyncMiddleware(handler))
   const postWithValidation = (path: string | string[], validationChain: ValidationChain[], handler: RequestHandler) =>
     router.post(path, ...validationChain, asyncMiddleware(handler))
 
@@ -23,6 +25,7 @@ export default function routes(services: Services): Router {
   )
   const cancelVisit = new CancelVisitController(services.auditService, services.visitService)
   const clearNotifications = new ClearNotificationsController(services.auditService, services.visitNotificationsService)
+  const confirmUpdate = new ConfirmUpdateController()
 
   // middleware to ensure valid visit reference
   // happens at this level to stop duplication
@@ -47,6 +50,9 @@ export default function routes(services: Services): Router {
     clearNotifications.validate(),
     clearNotifications.clearNotifications(),
   )
+
+  get('/:reference/update/confirm-update', confirmUpdate.viewConfirmUpdate())
+  post('/:reference/update/confirm-update', confirmUpdate.submitConfirmUpdate())
 
   return router
 }
